Create utility class for compareXml/compareTxt for refactoring purposes (#2209)
