+++ ieee-cdf/ieee-cdf-converter/src/test/java/com/powsybl/ieeecdf/converter/IeeeCdfImporterTest.java
@@ -6,6 +6,7 @@
  */
 package com.powsybl.ieeecdf.converter;
 
+import static com.powsybl.commons.ComparisonUtils.compareTxt;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
